+++ lib/data-migration.rb
@@ -3,7 +3,7 @@ require "data_migration/job"
 require "data_migration/task"
 
 module DataMigration
-  VERSION = "1.2.0".freeze
+  VERSION = "1.3.0".freeze
 
   module_function